+++ src/App.tsx
@@ -19,6 +19,13 @@ import PaymentCheckPage from "./routes/PaymentCheckPage";
 import PaymentResponsePage from "./routes/PaymentResponsePage";
 import CancelledPage from "./routes/CancelledPage";
 import KOPage from "./routes/KOPage";
+import { mixpanelInit } from "./utils/config/mixpanelHelperInit";
+
+declare const OneTrust: any;
+declare const OnetrustActiveGroups: string;
+const global = window as any;
+// target cookies (Mixpanel)
+const targCookiesGroup = "C0004";
 
 const checkoutTheme = createTheme({
   ...theme,
@@ -46,6 +53,27 @@ const checkoutTheme = createTheme({
 
 export function App() {
   const fixedFooterPages = ["payment", "qr-reader", "cancelled"];
+  React.useEffect(() => {
+    // OneTrust callback at first time
+    // eslint-disable-next-line functional/immutable-data
+    global.OptanonWrapper = function () {
+      OneTrust.OnConsentChanged(function () {
+        const activeGroups = OnetrustActiveGroups;
+        if (activeGroups.indexOf(targCookiesGroup) > -1) {
+          mixpanelInit();
+        }
+      });
+    };
+    // check mixpanel cookie consent in cookie
+    const OTCookieValue: string =
+      document.cookie
+        .split("; ")
+        .find((row) => row.startsWith("OptanonConsent=")) || "";
+    const checkValue = `${targCookiesGroup}%3A1`;
+    if (OTCookieValue.indexOf(checkValue) > -1) {
+      mixpanelInit();
+    }
+  }, []);
 
   return (
     <ThemeProvider theme={checkoutTheme}>
