+++ src/index.html
@@ -5,6 +5,7 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <title >React</title>
         <script src="env-config.js"></script>
+        <script src="onetrust.js"></script>
     </head>
     <body>
         <div id="root"></div>
